docs(form): separate each story into its own file
